fix(quota): symlink-safe chmod, sub-second ETA precision, error style
statefile.go: chmod the open temp-file descriptor (tmp.Chmod) instead
of the path (os.Chmod(tmpName)), which is vulnerable to a symlink
swap between create and chmod (CWE-377/CWE-59).

pace.go: convert ETASeconds to time.Duration by multiplying in
float64 space first; the previous int truncate-then-multiply order
zeroed any sub-second ETA.

admin_install_quota_schedule.go: stop bare _ = discarding the
launchctl unload error; check and ignore it explicitly with a reason,
per the repo style guide.

+++ pkg/llm/quota/pace.go
@@ -84,7 +84,7 @@ func parseCodexBarPace(data []byte) (map[string]*Pace, error) {
 				DeltaPercent:        rung.DeltaPercent,
 				ExpectedUsedPercent: rung.ExpectedUsedPercent,
 				WillLastToReset:     rung.WillLastToReset,
-				ExhaustsIn:          time.Duration(rung.ETASeconds) * time.Second,
+				ExhaustsIn:          time.Duration(rung.ETASeconds * float64(time.Second)),
 				Summary:             rung.Summary,
 			}
 			worst = morePressing(worst, candidate)
